+++ pkg/plugins/resources/k8s/consistent_kind_type_test.go
@@ -4,6 +4,7 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
+	"github.com/Kong/kuma/pkg/core/resources/apis/system"
 	core_registry "github.com/Kong/kuma/pkg/core/resources/registry"
 	k8s_registry "github.com/Kong/kuma/pkg/plugins/resources/k8s/native/pkg/registry"
 )
@@ -14,6 +15,10 @@ var _ = Describe("Consistent Kind Types", func() {
 		k8sTypes := k8s_registry.Global()
 
 		for _, typ := range types.ObjectTypes() {
+			if typ == system.SecretType {
+				continue // ignore Secret since we cannot map it directly to Secret K8S resource
+			}
+
 			res, err := types.NewObject(typ)
 			Expect(err).ToNot(HaveOccurred())
 			obj, err := k8sTypes.NewObject(res.GetSpec())
@@ -28,6 +33,10 @@ var _ = Describe("Consistent Kind Types", func() {
 		k8sTypes := k8s_registry.Global()
 
 		for _, typ := range types.ListTypes() {
+			if typ == system.SecretType {
+				continue // ignore Secret since we cannot map it directly to Secret K8S resource
+			}
+
 			res, err := types.NewObject(typ)
 			Expect(err).ToNot(HaveOccurred())
 			obj, err := k8sTypes.NewObject(res.GetSpec())