+++ tools/docs/kumactl/gen_help.sh
@@ -49,6 +49,8 @@ gen_help kumactl get traffic-permissions
 gen_help kumactl get traffic-routes
 gen_help kumactl get traffic-traces
 gen_help kumactl get fault-injections
+gen_help kumactl get secret
+gen_help kumactl get secrets
 gen_help kumactl delete
 gen_help kumactl inspect
 gen_help kumactl inspect dataplanes