+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## master
 * feat: pagination in kumactl
   [#690](https://github.com/Kong/kuma/pull/690)
+* feat: secret management API
+  [#684](https://github.com/Kong/kuma/pull/684)
 * chore: unify matching for TrafficPermission
   [#668](https://github.com/Kong/kuma/pull/668)
   ⚠️ warning: breaking change of matching mechanism